WIP: Sensu proof of concept for monitoring RDO things
This proof of concept piggybacks on the excellent work done by
Graeme Gillies, Emilien Macchi and other contributors to the
optools initiative around packaging Sensu and Uchiwa in Triple-O.

It leverages ansible to bootstrap the sensu master server and sensu
client servers and their basic configuration with the help of puppet
modules and then ansible takes over with a built-in core module to
manage sensu checks.

This is currently a work in progress.
What's missing and that I would like to add:
- Graphite integration within Uchiwa:
  http://roobert.github.io/2014/11/03/Sensu-with-Embedded-Graphite-Graphs/
- Grafana for viewing raw Graphite metrics
- More experimentation around checks (especially for Jenkins jobs)
- Manage iptables for further securing rabbitmq beyond ssl and authentication
- Tox jobs for testing these playbooks (ansible-lint and integration tests)

Change-Id: I2aaf7b2030eddc46deefeaba485b9f531acec543