Restructure tripleo-quickstart JJB
This is functionally a no-op, but just restructures the JJB for
the tripleo-quickstart jobs. This copies some of the structure
of the weirdo JJB. The promote templates/builders are split from
the gate templates/builders. This makes it a bit easier to add
periodic jobs for all gates in the next patch in this series.

Change-Id: I58f5525d5d7fb2bda0ab7f00ddba5bce06f5cfd2