Restructure tripleo-quickstart JJB
This is functionally a no-op, but just restructures the JJB
for the tripleo-quickstart jobs. This copies some of the structure
of the weirdo JJB.

The goal is to simplify and apply some DRY. Follow up patches
will add some periodic jobs, and are much simpler to reason
about because of this cleanup.

This requires a corresponding change to tripleo-quickstart:
https://review.openstack.org/#/c/320490

Change-Id: I58f5525d5d7fb2bda0ab7f00ddba5bce06f5cfd2