Add a PUBLISH variable to the get promote hash script
The only tripleo-quickstart job that is difficult to reproduce
outside of CI is the promote image building job. This is because
at the end of that job, it tries to upload images to the CentOS
artifacts server. We can control that behavior via a PUBLISH
variable, but in order to make sure it is only set to true when
running in CI, it is cleanest to set it in the promote hash file.

Change-Id: I84c5c4f983b8d5d8423413a302048e80951120ee