Reduce the number of jobs in the 'quick' queue
We doubled the amount of time the 'quick' jobs take, by moving
the default stopping point for quickstart.sh.

This patch reduces the number of jobs that run on each patch by
half, so that we get back to the same 'CI usage per patch' we had
before.

It runs the usbkey job on Mitaka, and the basic job on Liberty.
Any patches that have some release specific logic should run the
full-ci-check.

Change-Id: Id4eb274700cdce7e6bfbade5cb1fe20139bb1da8