Increase timeout for tripleo jobs.
We are seeing the gates for the image building role exceeding the 180
min timeout for a few different reasons.  While this is by no means a
long term change...we MUST get these times down, we're all-stop at the
moment without a bigger timeout.  Here's an example:

https://ci.centos.org/job/artib-gate-mitaka-delorean-full-images/46/

Let's unblock the jobs, then get the times down with intentional work.

Change-Id: I39ecb9d9440a72158707e48aa89d6f3270da0f60