Update patch set 1
Patch Set 1: Code-Review+1

This is a really good idea.  We just had a failure in the image gates that would have been caught if we were triggering jobs based on changes to images.sh.  So many different code paths run though this script that using it to gate (vs. relying on manually adding images-ci-check to jobs) is safer and an improvement.

Patch-set: 1