Increase defaults for quickstart jobs
We need the same change that was made here:

https://review.gerrithub.io/#/c/287168/

for the tripleo gate jobs as well.  The prior (above) change
was made to deal with the gates that are kicked when we need
to test changes made to the build role (artib).  These are the
converse gates, that are kicked when changes are made to oooq
where we think there might be build/image implications.

Here is a recent (just now) example of a build succeeding,
but failing due to the 180 min cap:

Job: https://ci.centos.org/job/tripleo-quickstart-gate-mitaka-delorean-full-images/61/console
Log: https://ci.centos.org/artifacts/rdo/jenkins-tripleo-quickstart-gate-mitaka-delorean-full-images-61/console.txt.gz

Spawned by: https://review.openstack.org/#/c/355633/

Change-Id: I60a36a10c8d40c72d498fdb50d2f0444a0354318
Related-to: https://review.gerrithub.io/#/c/287168/