update the location of the oooq images to match the delorean pin
Now that we will be publishing distinct images for RDO and the
production pipeline, we must publish the images to unique directories.
The $LOCATION variable defined in jobs/scripts/centos-$release is
passed to the image build role and to the publish job in the pipeline.

The naming convention used is based on which delorean softlink the
image was built from.  The promote-master job uses "consistent",
the master-current-tripleo promote job uses current-tripleo.

Change-Id: Ia1ef6d0cfa1b9489ba9fb94d3ef849f3966b3db5