Update patch set 1
Patch Set 1:

this is a bit confusing... this change to use a LOCATION var happened while I was out, and it is a bit hidden the way we are doing it.

In any case, we are changing that var from "testing" to "current-tripleo". Won't that change where we put the temporary image used by the promote job?

Patch-set: 1
Label: Verified=0