WIP: Handle publishing of testing images in JJB
This is currently handled via a "images/publish" role in
tripleo-quickstart, but really should not be, as it is very
specific to RDO image publishing.

This also adds a copy out to our recently provisioned file
server, so that we have a non-rate-limited copy to work with
outside of CI.

WIP: need to figure out how to pass variables to a publisher.

Change-Id: I1f5d5a5e59cde31e0d6bf6798568dc6df131741c