Update patch set 1
Patch Set 1:

I'm not going to block this but I would prefer if we found another solution.

What we are doing right now is silly:

    https://github.com/rdo-infra/ci-config/blob/master/jenkins/jobs/scripts/promote-upload-images.sh#L6-L10

We're downloading >3GB and re-uploading >3GB to the same destination except a slightly different directory.
Since artifacts uses rsyncd, it is not possible to do what would amounts to a "mv" instead.

However, it would be possible to instead maintain a symlink with rolling images. The promotion script would update the symlink to the proper location when a promotion occurs.

I guess, not unlike what we're doing with DLRN repositories right now.
We're definitely not downloading the whole repository and re-uploading it to current-passed-ci.

Patch-set: 1