Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

This is pretty close. Besides the comments inline, we also need to deal with the ssh key for images.rdoproject.org not being on the duffy nodes.

If that is going to block this though, we could just disable pushing testing images to rdoproject.org to at least get our promote job back running. That is a nice to have (albeit very nice), but not a must have.

Patch-set: 2
Label: Code-Review=-1