Add stable release testing with WeIRDO scenarios
This adds the release_rpm parameter to the WeIRDO generic jobs.
This also adds a new job that'll fire generic jobs with stable
releases. Only Mitaka is supported at this time but Newton will
be added once it is released.

The job is scheduled to run once every two days at midnight.

Change-Id: I59d379c440d1d55d6cecb18820d60f7bf06bbf5b