Update LOCATION and BUILD_SYS in cloudsig-testing jobs
We rely on the LOCATION variable in particular in the image consuming
jobs, in order to provide quickstart with the correct content.

The original patch reused "consistent" for location, even though this
does not make any since outside of a DLRN context. That also had the
side effect of clashing with the existing "consistent" release configs
in tripleo-quickstart.

This patch updates the image path to be: $RELEASE/cbs/cloudsig-testing

This allows us to create "cloudsig-testing" configs in quickstart which
point to the correct testing content:

https://review.openstack.org/#/c/381957/

Change-Id: I5ff537bae07434185b98ac58facde412b7645939