Remove the timeout from the weirdo wrapper/parent jobs
The new wrapper jobs are running into timeouts because the jenkins
slave is loaded and the child jobs, even if they are fast to complete,
are taking a while to actually be scheduled to run as the queue is
backed up.
Let's remove the timeout from those parent jobs since the child
jobs each have their timeouts configured properly.

At first glance, this patch adds wrappers, but this is to override
the default wrappers provided by the defaults which includes a 90
minutes timeout.

Change-Id: Ib4b5187a694a3f08f0bd8ed4511ae901957582c6