Test new image build process for master images
Our current image build process has been broken for some time.

First, we are using our own custom YAML for image building rather
than the packaged YAML from tripleo-common. This means anytime the
tripleo-common YAML is changed, we need to update the image build
role to match.

Second, we are polluting the host used to build images with DLRN
repos and openstack packages. We originally made a compromise on this
to limit the pollution which was to install a specific tripleo-common
package so as to not actually need DLRN repos on the build host. Again,
this causes drift from upstream, since we do not get any of the element
changes. Recently, this compromise finally broke for master. We
actually need a change in one of the elements in order to promote.

In order to solve the above problems without sacrificing the ability to
keep the build host clean, this patch introduces a method to build
images inside a CentOS guest image via virt-customize.
This way, we can just install the correct repos into the guest image,
then install tripleoclient and build images via the tripleoclient
command.

It then also converts the overcloud-full.qcow2 from the standard
process into an undercloud.qcow2. This undercloud.qcow2 can be used
with tripleo-quickstart in the same way as the one produced by the
current image build role.

This patch introduces a top-level ci-scripts directory to this repo
to house the new script. This directory can then also be used to
move some of the RDO CI specific scripts from tripleo-quickstart to
this repo.

In order to extend this work for older releases, we should probably
adapt the script into a new image build role in
openstack/tripleo-quickstart-extras that follows this new process.

Change-Id: I9366e01c9b123577a0b16be60d7ff3a07f7386cf