Update patch set 2
Patch Set 2:

(6 comments)

I like it!

A few questions inline.  Most of these can be addressed in subsequent commits, or the replacement role that this script can motivate creating.

I want to make sure we retain the ability to have a shared, parameterized mechanism driven in a declarative way (clear inputs), so we don't have "per release scripts" that end up having to be maintained individually.  I realize that this is more of a concern for downstream (OSP, RDO on RHEL, etc) than RDO master specifically, but it would be pretty nice if we could directly reuse this so that we can keep CI mechanisms in step for downstream where possible.

Nothing to give a hard -1 around

Patch-set: 2
Label: Code-Review=0