Update patch set 3
Patch Set 3: Workflow-1

I realized in testing the undercloud.qcow2 produced this way that it will require a tripleo-quickstart change as well.

This is because the image produced this way is no longer partitioned. I have a change almost ready that will handle both image types, but setting workflow -1 until I can get that submitted.

Patch-set: 3