Update patch set 3
Patch Set 3:

> why is that, aren't you using the same scripts as the previous image build role, just running them inside libguestfs VM?

Nope, this is a totally different process. The image build role today builds up all the images more or less in parallel. This why it needs to pre-install packages.

This new process instead builds the overcloud image exactly according to tripleo-common packaged YAML, with no pre-installed packages. Then it converts the overcloud-full to an undercloud image. The reason the undercloud created this way is not partitioned is because the overcloud full image is not partitioned.

Patch-set: 3
Label: Workflow=0