Use master image build method for current-tripleo pipeline
We are testing a new image building method in the master pipeline.
This same method can be used for the current-tripleo pipeline,
since it is just a different DLRN symlink on master. In fact, the
same issue that blocked the master pipeline will also block the
current-tripleo pipeline without this.

Change-Id: Ia64b2ae2b2f6c763be8763f2e84c72122315d247