remove upgrade ci jobs for ci.centos.org
Add discuss previoulsy in the upstream tripleo ci meeting
related to upgrade, we decide to drop the current upgrade
jobs (periodics) on centos-ci, since the workflow is too
big and the duration to long for the infra of centos.ci
The effort will be concentrate to the upstream tripleo-ci
jobs for Newton to Ocata (composable upgrade). This workflow
will be apply also in tripleo-quickstart to build composable
upgrade ci jobs on centos ci.
See for more informations:
https://etherpad.openstack.org/p/tripleo-upgrade-jobs

Note: the builder, job-template is not removed in this
review because it could re-used in a near futur for
implementing composable upgrade with tripleo-quickstart

Change-Id: I5a8e4a2b7e400e44b68532a8fac3e3e8a667cc2c