Collect log execution artifact
Presently, if the publisher calling the collect-logs role runs
without error, there will be no way of reviewing its exectuion.
This can be a problem because 1) we have no way of comparing a
known good execution vs a failed one and 2) the collect-logs role
itself is built to not fail if possible and subsequently issues
within CI can be hidden and incredibly difficult to debug without
locking up resources.

Change-Id: I14a1dc6cdb802b6ce7a306ffe8b4d0d5c1aa328c