Update patch set 2
Patch Set 2: Code-Review-2

This is not true. If the master release is inconsistent (there is a failure to build from source), the consistent hash will remain the same so long as the FTBFS is not resolved and this can sometimes - unfortunately - last for some time.

Can you find another way to address this requirement for the tripleo pipeline ?

Patch-set: 2
Label: Code-Review=-2