Update image building script to use build-images role
This updates the master job image builder to use the new
build-images role in tripleo-quickstart-extras. It replaces the
fake ansible script we have that was used to test the methodology
which became the new build-images role.

This is also the first ci-script which will live in this repo
rather than the ci-scripts directory in tripleo-quickstart. We have
been asked to move all of the ci.centos related ci scripts from
the upstream quickstart repo. As such, this will prove the approach.

We do not put the script in jjb, but rather in a top level ci-scripts
directory. This will allow us to setup a gate in jjb that runs on
patches to this repo that change the ci-scripts. It is not possible
to gate scripts which are directly in the jjb in this way.

**NOTE**
 - checkout-patches.sh is temporary to include the unmerged quickstart
   patches for the new build-images role. We are under pressure to get
   the Ocata pipeline setup, and this is blocking that effort. The
   approach in the patches included is agreed upon, so they are very
   low risk.

Change-Id: I6a487386d8d9cb0d1d46e3b9cf0b95bc5758d587