WIP Update image building script to use build-images role
This updates the master job image builder to use the new
build-images role in tripleo-quickstart-extras. It replaces the
fake ansible script we have that was used to test the methodology
which became the new build-images role.

This is also the first ci-script which will live in this repo
rather than the ci-scripts directory in tripleo-quickstart. We have
been asked to move all of the ci.centos related ci scripts from
the upstream quickstart repo. As such, this will prove the approach.

We do not put the script in jjb, but rather in a top level ci-scripts
directory. This will allow us to setup a gate in jjb that runs on
patches to this repo that change the ci-scripts. It is not possible
to gate scripts which are directly in the jjb in this way.

**WIP**
 - the only thing left WIP about this is the checkout-patches.sh
   script. Ideally, we merge the build-images role patches and this
   is not needed. I included it in case we are blocked and want/need
   to merge this sooner.

Change-Id: I6a487386d8d9cb0d1d46e3b9cf0b95bc5758d587