Dont fail publish if missing artifacts
We don't want to fail the publish if we are missing
files in the artifact_list. If we are somehow missing
an undercloud image at this point, the subsequent test
jobs will fail. That is preferable to blocking the jobs
which do not even use that image from running though.

Change-Id: I3e5f0b447a9224a15e351c5506bc68225e0e0ef1