Cherry pick send mail tool to test in full tempest
This change gets the change still under review in
https://review.openstack.org/#/c/423340/ and apply in
tripleo-quickstart-extras, so we can test the full tempest job sending
email with the results, as a POC before we send it to upstream.
Please notice this will not add a new job on centos ci, it will just use
the actual full-tempest job that runs daily, so will not overload the
ci.

Change-Id: I61607883dcd33b5153cf12cdd4b43bb8804a855c