Update patch set 1
Patch Set 1: Code-Review-1

(4 comments)

Was waiting for the build-images role to be merged before setting this up, but we could do it before so we could potentially gate the patch adding it. 

The only tricky bit there is that we have a script manually checking out patches which would override the gated change. I put up a review to remove that though: https://review.rdoproject.org/r/#/c/5091/

I will rebase that patch on this one, once we update it to use the rdo-infra builder.

Patch-set: 1
Label: Code-Review=-1