Make containers gate able to depend on TripleO patches
The containers gate was not set up as a 'dlrn-gate-check' job_type.
That is the job_type which actually runs the build-test-packages
role, so we need it in order to test TripleO patches via Depends-On.

This patch also moves the job into the project block with the other
tripleo-quickstart-extras libvirt jobs.

Change-Id: Iba4bedcfd0a11614b5e6f2385ff5fef35867c700