Move weirdo generic jobs from duffy to openstack virtual machines
We're having a hard time with bare metal hardware capacity in ci.c.o.
This patch introduces a builder and a publisher to respectively
create and destroy virtual machines for the purpose of jobs.

Since WeIRDO already expects an inventory file in $WORKSPACE/hosts,
this makes no difference to the jobs at all, we're just changing
how we generate the hosts file and the target is now a VM instead
of a bare metal.

Considerations:
- ~/.config/openstack/clouds.yaml exists with the right credentials
  on every jenkins slave
- The keypair is set up on the target cloud
- The image is set up on the target cloud
- There is a log server set up with pubkey authentication for
  the "uploader" user

Change-Id: If4b350fd9eee852a91e0a3df3137b23009c7a2db