Move weirdo generic jobs from duffy to openstack virtual machines
We're having a hard time with bare metal hardware capacity in ci.c.o.
This patch introduces a builder and a publisher to respectively
create and destroy virtual machines for the purpose of jobs.

Since WeIRDO already expects an inventory file in $WORKSPACE/hosts,
this makes no difference to the jobs at all, we're just changing
how we generate the hosts file and the target is now a VM instead
of a bare metal.

Considerations:
- ~/.config/openstack/clouds.yaml exists with the right credentials
  on a given jenkins slave
- A DIB image built with the "weirdo-dib.sh" script is set up on
  the target cloud with the name "template-centos7-weirdo"
- There is a logserver "logs.rdoproject.org" available for pulling
  logs from the provisioned virtual machine and for pushing the ara
  report from the jenkins slave

Change-Id: If4b350fd9eee852a91e0a3df3137b23009c7a2db