Update quickstart-extras patch for build-images role
We have a bit of a chicken egg problem with setting up a gate for
the build-images role before having the build-images role merged.

We will use this to make sure the latest patchset does not break
our promotion job before merging the build-images role, and then
removing this checked out patch via:

https://review.rdoproject.org/r/5091

Lucky number 27?

Change-Id: I49ebe288d26bc5d10bb6358d631e5dc10cf90891