Use master-tripleo-ci gate check on the full gate job
This makes the "full-ci-check" work on the same configuration as master
gate on tripleo-quickstart-extras.

Previously the quickstart master gate was not using the code that checks
the build-test-packages role which might be broken by quickstart
changes. It will also use the master-tripleo-ci release that exercises
more parts of the tripleo-quickstart code.

Change-Id: Ib8715c80e03978a643212484787f5858d3a45ffc