Remove rdo-delorean-promote-master pipeline
https://review.rdoproject.org/r/#/c/5610/ added weirdo
jobs to rdo_trunk-promote-master-current-tripleo to add
full coverage to current-tripleo-rdo.

This patch removes rdo-delorean-promote-master to have a
single pipeline for master repos and use the resources in
ci.centos.org in a efficient way. Once this is merged
and get current-tripleo-rdo ahead of current-passed-ci
we'll set current-passed-ci as a symlink to current-tripleo.rdo.

Change-Id: If4508ae4e6f1d4434ff24da064859287ad926634