Update patch set 1
Patch Set 1: Code-Review-1

I hate to be that guy but can we re-open this for discussion ?

I feel like the main purpose of trying to consolidate into the TripleO pipeline was a compromise to help us optimize the amount of resources we used from ci.centos, especially because we needed to support 3 stable releases due to the Ocata short cycle.

Where we stand today:
- Mitaka is going EOL next week, meaning a bunch of gate jobs and promotion jobs will disappear
- We've started moving weirdo jobs to rdo-cloud (generic jobs only, promotion jobs once enough confidence is built)
- TripleO-Quickstart continues their effort in migrating their gate jobs upstream

The scope of the tripleo pipeline, as I understand it, is to provide additional coverage on top of what tripleo-ci covers and to provide a lower "barrier to entry" for promoting content for downstream import purposes.

That's why it will only run on tripleo-ci "vetted" hashes and that's the problem I see in removing potential visibility and insight into problems beyond what tripleo-ci is currently standing on.
If there are no tripleo-ci promotions for a while, this means we have no visibility on other problems - either puppet-openstack or packstack specific - or something that TripleO simply does not test.

If someone brings the argument of "confusion" between the pipelines,  I've written about this before [1]. Jenkins has a description field for jobs and we're indeed doing a terrible job at explaining what the jobs are for. Let's improve that instead.

[1]: https://www.redhat.com/archives/rdo-list/2017-January/msg00037.html

Patch-set: 1
Reviewer: Gerrit User 16 <16@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1