Update patch set 1
Patch Set 1:

Can you expand on what exactly do you need ?
The entirety of the job is an ansible playbook with a role, there's nothing else [1].

All the containers are built in separate tasks, explicitely for the purpose of making it easier to troubleshoot because troubleshooting kolla-build sucks since it parallelizes builds and interleaves the output of the builds in one single stream.

Here's a failure example:
    https://logs.rdoproject.org/rdo-tripleo-containers-build/14/ara/
    https://logs.rdoproject.org/rdo-tripleo-containers-build/14/ara/result/3bc434b6-a8ee-4908-9f1e-88403bd53771/

And another:

    https://logs.rdoproject.org/poc-dmsimard-containers-push/2/ara/
    https://logs.rdoproject.org/poc-dmsimard-containers-push/2/ara/result/3b2eede3-9b35-410a-86ab-8ad0ef3446e3/

Host facts are available as well: https://logs.rdoproject.org/rdo-tripleo-containers-build/15/ara/host/366f7714-c11e-4059-b6fb-f5caa7a1600a/

It's entirely fair if you are not familiar with ARA but we can address that.

[1]: https://github.com/rdo-infra/ci-config/blob/master/jenkins/jobs/tripleo-containers.yml#L2-L68

Patch-set: 1