Update patch set 1
Patch Set 1: -Code-Review

@David, I like your idea here https://gist.github.com/dmsimard/0b6e1720b7fa1f7d0f696fc00b6418ab.

Having something that highlights the error is great. I found myself scrolling through tasks search for the failure.  I tried sorting but the failure did not bubble to the top.

After looking through build 14 w/ the failed example it's more clear now.  I will help newbies if you include the feature in this gist, but lets move forward and iterate on it.   

@Alan the logging here is not bad, just different.  I suspect including a console.log along w/ ara may make this look at feel more standard. Including the feature in the gist is also a nice add.

Patch-set: 1