Cleanup stale VMs in cico jobs
Sometimes, VMs fail to start or stop for some reason and are left
in an error state in the RDO Cloud. This consumes our quota, and
needs to be cleaned up manually to avoid issues with later jobs.

This commit adds an automatic cleanup feature, so before every
VM creation we check for any VM in "ERROR" state or older than
6 hours (21600 seconds), and remove it.

Change-Id: I7494ce834add9359e05d341380458fc294e1350d