Update patch set 1
Patch Set 1:

Cleaning up VMs in error state is good but we should also be checking for orphaned virtual machines -- i.e, that have not been properly deleted for a reason or another.

We could do that based on the date of creation of the virtual machine -- any VM should not be up for more than, let's say, 6 hours.

Patch-set: 1
Label: Workflow=0