Update patch set 3
Patch Set 3:

weirdo actually does use fact caching [1] but in hindsight it's not necessary -- we don't really run a series of playbooks against multiple hosts which would make that cache useful.

Since it's not useful, I'm okay for disabling it -- I'll send a patch to remove it from weirdo.

[1]: https://github.com/rdo-infra/weirdo/blob/master/playbooks/ansible.cfg#L5-L8

Patch-set: 3