Fix stale VM cleanup playbook
The previous version of the playbook used to cleanup stale VMs had
an issue: when the playbook date (gathered as a fact at the beginning
of the execution) was lower than a VM creation date, the number of
seconds in the timedelta object was not negative, but positive and
quite high. That caused a condition where a VM that was created shortly
after the playbook was started would be removed.

Fixing this by using the days property, and only cleaning up VMs that
are older than 1 day. The days property *can* be negative, so a VM
created after the playbook was started will give us a value of "-1 days"
in the substraction, and won't be removed.

This reverts commit c9e8264d360b9377174140582d332bedc95acba0.

Change-Id: Iedce106f1ddef0388240c8ed6132e4c08260a825