Skip p-o-i scenario004 in generic jobs for newton
Currently, scenario004 has been added to weirdo-testing and
weirdo-release templates. since then, newton cloudsig-testing
and cloudsig-stable pipelines are failing as this job is not
not compatible with newton.

This patch makes the weirdo generic builder to skip scenario004
for newton. Although the optimal solution would be not adding
the job for pike pipelines, the only solution i could find is
to create separated templates for newton which would add
complexity to JJB, so i rather this solution.

Change-Id: I31988bb1072a7ca83284f362389766811fcfe9ef