Update patch set 2
Patch Set 2:

So just making sure I understand here, you're ultimately just changing an exit 1 for an exit 0 right ?

Also the other thing with the compat check builder is that it ran before even provisioning the virtual machine so we were not spinning up a VM for nothing.

The problem is that if we exit 0 in the compat check builder, it doesn't exit the job, it considers the builder successful and keeps going.

Patch-set: 2