Simplify VM name to prevent SSL CN creation issues
Hostnames in weirdo VMs are becoming too long, and we get some names
like 'weirdo-master-promote-packstack-scenario001-1161.openstacklocal'
which are 64-chars long. This is creating an issue in some Packstack
jobs, because there is a limitation of 64 characters in the CN used
by a X.509 certificate.

This patch simplifies the hostname by removing the 'weirdo' prefix
and the scenario keyword, leaving just the scenario number.

Change-Id: I4e4bfba4f7f0c54ff77c09a31c107b9f06a3d342