Update patch set 5
Patch Set 5:

> Just for the record, you can use an Ansible task directly to report
 > if you don't want to use a shell script, see [1].

The requirement is to be a post-build shell script, so it's easier to use the command line client. But I thought of that too, thanks for pointing it out.

 > Also, I don't see any place where the dlrnapi_client module is
 > installed.

I added it to quickstart's requirements.txt and I'm using it from the virtenv, but maybe it would be more simple to create a one-off virtenv here for reporting.

 > [1] - https://github.com/rdo-infra/weirdo/blob/master/playbooks/dlrn-api-report.yml

Patch-set: 5
Label: Code-Review=0