Update patch set 3
Patch Set 3:

(6 comments)

> (7 comments)
 > 
 > I have some comments on the documentation and a small suggestion,
 > otherwise it looks fine.
 > 
 > My only (small) concern is that we're using the dlrnapi_client
 > methods directly instead of its command-line, but it should be
 > stable enough and not change often.

I thought that was an intended use of the dlrnapi_client as you have an example like this: https://github.com/softwarefactory-project/dlrnapi_client/#getting-started

So I assumed it's normal to interface with the library itself. Why I also chose to do this is because the pprint()ed output of the script is not too friendly to process... It's not JSON nor YAML, so I can't reall process it easily. I was thinking about how to do this best but realized the easiest is to use the python library.

Patch-set: 3