Update patch set 3
Patch Set 3:

> I'd prefer to have dry_run as a command line argument instead of a
 > configuration option

I agree that it's a bit unconventional that dry_run is not a command line option, but this way we can enable/disable the whole pipeline with changes to this file instead of changing JJBs/builder scripts at different places.

I intentionally didn't add an argparse part to discourage this type of use. I think this is the right approach for a CI tool like this, but I can be convinced otherwise. :)

Patch-set: 3
Label: Code-Review=0