Update patch set 2
Patch Set 2:

(1 comment)

I dont think we should touch the jjb scripts, because those will just be obsolete once we have image promotion outside of CI with the rest of the promote process.

As far as the upload script, I think it is correct wrt the card to bring everything in line. However, backwards compat will be an issue.

We are trying to change what is currently a directory to become a symlink:

https://images.rdoproject.org/master/rdo_trunk/

I am not sure how to do that without deleting the directory and breaking anyone using those links.

Patch-set: 2
Reviewer: Gerrit User 84 <84@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0