WIP: figure out the last good hash
This is a quick and dirty script that will go back in the promotion
history and tries to find the last hash where the promotion criteria as
true.

Run it just like the promoter script, i.e.

    python last-good.py config/pike.ini

The log will be printed on the console. There are log messages and
function names with "promote" in them, but in fact this script will
never try to promote anything, it is read only.

Change-Id: Iefc32873c5fd33e56fc754752ec7ffeb2de37973