Update patch set 3
Patch Set 3:

Attila, John, Wes, I'm not convinced we need a dedicated machine for this.
Can we take some time to discuss if a cron on a machine is the right approach in the first place ?

I feel there's a lot of different options available but we haven't had the opportunity to discuss them.

For example, a jenkins job that would trigger periodically or through content change on the DLRN API result pages ? I'm sure we could come up with different ideas.

Patch-set: 3
Reviewer: Gerrit User 16 <16@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0