Promote repositories in both trunk.r.o and trunk-primary.r.o
In RDO there are two DLRN servers, trunk-primary.rdoproject.org,
which is used to build the repos and trunk.rdoproject.org which
serves the repos. We want to promote the repos in both servers
in order to be consistent in case we want to use trunk-primary
as public server in public server outage.

Change-Id: I7520054895df176d0c608e36295d405484356994