Update patch set 2
Patch Set 2:

From an API point of view, when promoting just meant "adding a symlink locally" it probably made sense to promote both internally and externally.

Now that promoting also involves a DB access, it is a better idea to only promote once, in the public instance, and then use that one for everything. Using trunk-primary doesn't really buy us a lot, since we're not in the situation we were when it was inside the ci.centos network.

Patch-set: 2
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0