Add a script for the uploader jobs to create an undercloud image
Since implementing the new promotion pipeline we are no longer
creating an undercloud image. This is fine for quickstart users,
because we have the ability to use an overcloud-full image as an
undercloud.

However, RDO has published an undercloud image for almost 2 years
now, so there are other projects that depend on that image existing.

We can solve that issue by running the same conversion process we
do on the fly in quickstart at the end of the uploader jobs. Then,
uploading this converted undercloud along with the overcloud
artifacts.

The RDO promote pipelines will get the added benefit of not needing
any image build job at the start, as they can just download the
appropriate undercloud image for the delorean hash they are testing.
(They could do the same with the overcloud images, but we need at
least some validation of this undercloud image.)

The uploader jjb is in the review.rdoproject.org config repo, so
we will need another patch there in order to use this script.

Change-Id: Id476eb50534ef0e2b3014dd94f65374f808bb8b1