Revert "Revert "Move weirdo jobs in ci.centos to ci aggregate""
Jobs are regularly timing out and perforance problems have been
observed in servers in RDO as review.rdoproject.org. Let's move
jobs back to SSDs and check the impact.

Note that the number of jobs should be too high, as it's mainly
promotion pipelines and jobs to gate CBS promotions only.

This reverts commit c3c9b27ad282dca25235220ab28cdaad3ba1448b.

Change-Id: I428718573c534418c2037de8be9a62a1a17b392a