Promoter: promote hash only after other steps completed
This change promotes the hash only after we're sure all the other steps
in a promotion (cloud images tagging, container images upload) have
completed successfully.
Until now, promoting before everything was guaranteeing that no two
promotion script run attempted to promote the same hash. It simply
discovered that the hash was already promoted, and exited. Now that we
are doing it last, we need to ensure that a second run will not attempt
while the first is still completing all the steps, hence the locking
mechanism

Change-Id: I490dabac4380d3ab1423706f4a9ce63e88501ebf