Promoter: check a list of the 5 latest hash for a valid candidate
Until now, the promoter script checks only the latest hash pointed at by
specific name. If one of the previous candidate is skipped for some
reason and it could be a valid candidate, it's not considered in any
way.
This could lead to a situation when an hash that could be promoted does
not get promoted, and the very latest hash is not a valid candidate, and
we don't get any promotion.
This change expands the logic to consider a history of the latest 5
hashes, if the latest cannot get promoted, check the second latest, and
so on.

Change-Id: I6fd162ca480b8e9aa42a62fe33f4f60b2ce255d5