Add timeout to promoter script run
After we implemented a locking mechanism to prevent simultaneous
promotion runs, we can end up with the promotion process hung.

When this happens, someone needs to manually log on to the server
and kill the process.

This patch adds a timeout to the promote run, so that it will no
longer require manual intervention.

Closes-Bug: 1737617
Change-Id: I18e0509d6d3faa686d2d8851ff9a1aac6aa93665