Use container list from tripleo-common commit being promoted
We are getting the list of containers to push dynamically, by
downloading it from the tripleo-common git repo. Currently, we
are just specifying the release in order to get the correct
container list for the particular release.

This has a flaw in that the list of containers can be updated in
the time it takes for all phases to promote. When this happens,
the phase that has not yet promoted will get stuck, because it can
not find the new containers.

This change addresses the issue by using the exact commit from
tripleo-common to download the list of containers. This ensures
that the list is the same through all phases of the promote
pipeline, since the tripleo-common commit being promoted will be
the same throughout.

Change-Id: Ie026210aabc1c9398401d890842b7dcb0c159396
Closes-Bug: 1746056