Only check if lastest_hashes are newer if we have old_hashes
Currently we try to check the timestamp of old_hashes even if we
have no old_hashes. This leads to an exception and failure of the
script in the case a release has never been promoted. Instead,
we only check the newness of latest_hashes if we have old_hashes
to compare to.

Change-Id: I2e62d25332a1a5276a6a278fc71f0c2cc33dab2e