Log exceptions in promoter script to web logs
Currently if an uncaught exception happens in the promoter script,
as in the case of a new bug, the failure is only logged to the
cron logs in /var/spool/mail. This makes debugging impossible
without logging into the promote server. Instead, we can wrap the
entire promoter call in a generic try/except and log any exceptions
to the web logs.

Change-Id: Ie85dd97e8cfbaf535c03e6543a3c336eeee68999