Update grep for tripleo-common in dynamic container list
With the addition of tripleo-common tempest plugin, there are now
2 packages in tripleo-common in the name.

This patch updates the grep to get only the hash of the
openstack-tripleo-common package, and should be resilient to a
future package with tripleo-common in the name.

It also adds a check for the container list being empty and fails
the playbook in this case, since we do not want to proceed
promotion without containers pushed.

Change-Id: I6c8643aa47276049cc8621a94bb49a7168071ad2