Update patch set 1
Patch Set 1: Workflow+1 -Code-Review

as we discussed, the name can be anything, but we need to create the keypair  in the playbook and it needs to contain all the keys of tripleo ci team and rdo infra in it. We'll do it only once, so it will work for every run.
We could merge this now, but the key part will need to be addressed for the card to be completed.

Patch-set: 1
Reviewer: Gerrit User 63 <63@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=+1