Remove minimal_pacemaker tripleo jobs from RDO phase 1 criteria
The tripleo jobs in RDO phase 1 are really only adding coverage of
libvirt based quickstart at this point. As such, and given we
consistently run out of nodes in ci.centos, it makes sense to just
run a single job to validate that the libvirt path is not broken.

This patch also removes the minimal_pacemaker gates from the
quickstart repos, because we do not want to have gates that are
different than what we have in the promote criteria.

Change-Id: I993ae0c91fd5e8f40d0fbef4b8fc068222320d6f