Update patch set 8
Patch Set 8:

@apevec so afaict David's suggestion to follow the pattern with a bastion host where the infrastructure was deployed from a single host is a good one and one we intend to get to.  

Given the remaining time in this infrastructure sprint we determined w/o a proper bastion and automatic deployment and provisioning of the environment a "jump host" was an additional layer of security such that the ssh port on a public ip would only be available on one server.

David's comments do not imply a jump host is a bad thing to have, however it's not the ideal end state which we understand and will develop at a later time.

Patch-set: 8