dlrnapi promoter: raise all ApiExceptions
Presently there are 2 cases where the promoter script
will swallow ApiExceptions (bad) and return 'None' vs.
invoking a 'raise' instruction.  This causes difficult
to debug issues such as:

- https://bugs.launchpad.net/tripleo/+bug/1756892

In this specific case the script created a "new" promotion
event via the API for an older hash, which gives it a new,
current timestamp.

Change-Id: I4d500a219810c163fd999f6e7e8265aa5de63120
Related-To: Bug #1756892