Fix missing config_file variable
In moving the config parsing into the main function in
dc8a4d63b29b1818c7cc2676748320ff28722155 we missed the fact that
config_file was used in the promoter function to log what config
file was being used.

This explodes because the config_file variable is no longer in
scope for that function.

Instead, we can just log the release here.

Change-Id: I209fb13ef05e9dd9b6c88d24a322f3baccaaa484